Fix warning in Spells.cpp: change variable to size_t.

+++ src/game/Spells.cpp
@@ -5557,8 +5557,7 @@ bool ARX_SPELLS_Launch( const long& typ, const long& source, const long& flagss,
 				
 				SPELLCAST_Notify(i);
 
-				for (long n=0;n<MAX_SPELLS;n++)
-				{
+				for(size_t n = 0; n < MAX_SPELLS; n++) {
 					if (!spells[n].exist) continue;
 
 					switch (spells[n].type)
